check if text is correctly decyphered

+++ check_if_cracked.h
@@ -4,5 +4,5 @@
 
 #ifndef CHECK_IF_CRACKED_H
 #define CHECK_IF_CRACKED_H
-int check_if_cracked(char *text);
+float check_if_cracked(char *text);
 #endif //CHECK_IF_CRACKED_H
